feat: DAM-25 UseCase 구조 변경
