--- domain/src/main/java/com/khnsoft/damta/domain/request/Request.kt
@@ -1,3 +1,0 @@
-package com.khnsoft.damta.domain.request
-
-interface Request
