--- domain/src/main/java/com/khnsoft/damta/domain/request/place/SearchPlaceRequest.kt
@@ -1,10 +1,0 @@
-package com.khnsoft.damta.domain.request.place
-
-import com.khnsoft.damta.domain.request.Request
-
-data class SearchPlaceRequest(
-    val keyword: String,
-    val page: Int,
-    val size: Int,
-    val isAddress: Boolean = false
-) : Request
