--- domain/src/main/java/com/khnsoft/damta/domain/response/EmptyResponse.kt
@@ -1,3 +1,0 @@
-package com.khnsoft.damta.domain.response
-
-object EmptyResponse : Response
