--- domain/src/main/java/com/khnsoft/damta/domain/response/Response.kt
@@ -1,3 +1,0 @@
-package com.khnsoft.damta.domain.response
-
-interface Response
