--- domain/src/main/java/com/khnsoft/damta/domain/response/area/AddAreaResponse.kt
@@ -1,7 +1,0 @@
-package com.khnsoft.damta.domain.response.area
-
-import com.khnsoft.damta.domain.response.Response
-
-data class AddAreaResponse(
-    val areaId: Int
-) : Response
