--- domain/src/main/java/com/khnsoft/damta/domain/response/place/SearchPlaceResponse.kt
@@ -1,8 +1,0 @@
-package com.khnsoft.damta.domain.response.place
-
-import com.khnsoft.damta.domain.model.Place
-import com.khnsoft.damta.domain.response.Response
-
-data class SearchPlaceResponse(
-    val placeList: List<Place>
-) : Response
