--- domain/src/main/java/com/khnsoft/damta/domain/response/user/UpdateUserResponse.kt
@@ -1,8 +1,0 @@
-package com.khnsoft.damta.domain.response.user
-
-import com.khnsoft.damta.domain.model.User
-import com.khnsoft.damta.domain.response.Response
-
-data class UpdateUserResponse(
-    val user: User
-) : Response
