+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/Request.kt
@@ -1,0 +1,6 @@
+package com.khnsoft.damta.domain.usecase
+
+interface Request {
+
+    object Empty : Request
+}
