+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/Response.kt
@@ -1,0 +1,6 @@
+package com.khnsoft.damta.domain.usecase
+
+interface Response {
+
+    object Empty : Response
+}
