+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/ResultUseCase.kt
@@ -1,8 +1,5 @@
 package com.khnsoft.damta.domain.usecase
 
-import com.khnsoft.damta.domain.request.Request
-import com.khnsoft.damta.domain.response.Response
-
 interface ResultUseCase<R : Request, T : Response> {
 
     suspend operator fun invoke(request: R): Result<T>
