+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/user/SignInUseCase.kt
@@ -3,21 +3,26 @@ package com.khnsoft.damta.domain.usecase.user
 import com.khnsoft.damta.common.extension.flatMap
 import com.khnsoft.damta.domain.common.HashGenerator
 import com.khnsoft.damta.domain.repository.UserRepository
-import com.khnsoft.damta.domain.request.user.SignInRequest
-import com.khnsoft.damta.domain.response.EmptyResponse
+import com.khnsoft.damta.domain.usecase.Request
+import com.khnsoft.damta.domain.usecase.Response
 import com.khnsoft.damta.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
-internal class SignInUseCase @Inject constructor(
+class SignInUseCase @Inject constructor(
     private val userRepository: UserRepository,
     private val hashGenerator: HashGenerator
-) : ResultUseCase<SignInRequest, EmptyResponse> {
+) : ResultUseCase<SignInUseCase.SignInRequest, Response.Empty> {
 
-    override suspend fun invoke(request: SignInRequest): Result<EmptyResponse> {
+    override suspend fun invoke(request: SignInRequest): Result<Response.Empty> {
         val hashedPassword = hashGenerator.hashPasswordWithSalt(request.password, request.username)
 
         return userRepository.signIn(request.username, hashedPassword).flatMap {
-            Result.success(EmptyResponse)
+            Result.success(Response.Empty)
         }
     }
+
+    data class SignInRequest(
+        val username: String,
+        val password: String
+    ) : Request
 }
