+++ domain/src/main/java/com/khnsoft/damta/domain/usecase/user/UpdateUserUseCase.kt
@@ -3,16 +3,18 @@ package com.khnsoft.damta.domain.usecase.user
 import com.khnsoft.damta.common.extension.flatMap
 import com.khnsoft.damta.domain.common.UserValidator
 import com.khnsoft.damta.domain.error.UserError
+import com.khnsoft.damta.domain.model.User
 import com.khnsoft.damta.domain.repository.UserRepository
-import com.khnsoft.damta.domain.request.user.UpdateUserRequest
-import com.khnsoft.damta.domain.response.user.UpdateUserResponse
+import com.khnsoft.damta.domain.usecase.Request
+import com.khnsoft.damta.domain.usecase.Response
 import com.khnsoft.damta.domain.usecase.ResultUseCase
+import java.time.LocalDate
 import javax.inject.Inject
 
-internal class UpdateUserUseCase @Inject constructor(
+class UpdateUserUseCase @Inject constructor(
     private val userRepository: UserRepository,
     private val validator: UserValidator
-) : ResultUseCase<UpdateUserRequest, UpdateUserResponse> {
+) : ResultUseCase<UpdateUserUseCase.UpdateUserRequest, UpdateUserUseCase.UpdateUserResponse> {
 
     override suspend fun invoke(request: UpdateUserRequest): Result<UpdateUserResponse> {
         when {
@@ -32,4 +34,14 @@ internal class UpdateUserUseCase @Inject constructor(
             Result.success(UpdateUserResponse(user))
         }
     }
+
+    data class UpdateUserRequest(
+        val nickname: String,
+        val email: String,
+        val birthday: LocalDate
+    ) : Request
+
+    data class UpdateUserResponse(
+        val user: User
+    ) : Response
 }
